Advance database schema to version 2
Made two changes to the database schema:
1. Added "Closed" field to Accounts.
2. Added "Fee" field to Transactions.

Continued to add code to support the purchase transaction. Cost basis
is now able to be computed thanks to the inclusion of "Fee", which is
subtracted from "Value".

Added a field to the purchase dialog for the name of the symbol. This
is used when the symbol isn't already in Securities. This is a bit
clunky but it will do for now. I am not sure what better way there is
to add a new security; I am not confident in having the app go to an
online source to get info on the symbol.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>