Finish PurchaseTransactionDialog
Added code to validate the symbol that the user entered. The rule
is now that it must be 1-5 upper-case letters. If not valid, an
ErrorProvider is used to direct the user's attention to the error.

Thus, all symbols in the database will be in upper case.

At this point, recording a purchase is complete.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>