Various fixes
Fixed bug where the schema version hadn't been updated to 3, which allowed
schema 2 files to be opened.

Fixed bug where the commission was not being subtracted from the transaction
cost when adding the price to the database.

Changed the name of GetSecurityPrice to GetPrice to make consistent with
the naming scheme.

Added Selecting event handler for the tab control so that it will
specify the first account in the combo box, if it is not already
set. Before, the whole tab was blank, forcing the user to select
an account manually.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>