GetAccountNameList() fix and unit tests
GetAccountNameList() did not handle the "no accounts" case correctly.
In fixing this bug, I decided to add a couple of unit tests.

The unit tests as a whole were also refactored to use a DB set up
automatically for each test, via [TestInitialize]. This DB has accounts,
securities, and prices that are used by most of the unit tests.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>