Implement Deposit Cash transaction type
This is the initial implementation of the deposit cash transaction type.
The menu item is active, and when selected the Deposit Cash dialog is
shown. After the user enters the info and clicks OK, the transaction is
added to the database and the file is saved. The transaction view is
updated with the new info.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>